exit syscall added

+++ include/stdlib.h
@@ -4,7 +4,7 @@
 #include <sys/defs.h>
 
 int main(int argc, char *argv[], char *envp[]);
-void exit(int status);
+void exit(uint64_t status);
 
 void *malloc(size_t size);
 //void free(void *ptr);
